Crate reth_ethereum_forks

Ethereum fork types used in reth.

This crate contains Ethereum fork types and helper functions.

Feature Flags

arbitrary: Adds proptest and arbitrary support for primitive types.

Structs

ForkFilter

Filter that describes the state of blockchain and can be used to check incoming ForkIds for compatibility.

ForkHash

CRC32 hash of all previous forks starting from genesis block.

ForkId

A fork identifier as defined by EIP-2124. Serves as the chain compatibility identifier.

ForkTransition

Represents a transition from one fork to another

Head

Describes the current head block.

Enums

ForkFilterKey

How to filter forks.

Hardfork

The name of an Ethereum hardfork.

ValidationError

Reason for rejecting provided ForkId.