Enum reth_ethereum_forks::Hardfork

#[non_exhaustive]
pub enum Hardfork {

    Frontier,
    Homestead,
    Dao,
    Tangerine,
    SpuriousDragon,
    Byzantium,
    Constantinople,
    Petersburg,
    Istanbul,
    MuirGlacier,
    Berlin,
    London,
    ArrowGlacier,
    GrayGlacier,
    Paris,
    Shanghai,
    Cancun,

}

The name of an Ethereum hardfork.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Frontier

Frontier.

Homestead

Homestead.

Dao

The DAO fork.

Tangerine

Tangerine.

SpuriousDragon

Spurious Dragon.

Byzantium

Byzantium.

Constantinople

Constantinople.

Petersburg

Petersburg.

Istanbul

Istanbul.

MuirGlacier

Muir Glacier.

Berlin

Berlin.

London

London.

ArrowGlacier

Arrow Glacier.

GrayGlacier

Gray Glacier.

Paris

Paris.

Shanghai

Shanghai.

Cancun

Cancun.

Implementations

impl Hardfork

pub fn mainnet_activation_block(&self, chain: Chain) -> Option<u64>

Retrieves the activation block for the specified hardfork on the Ethereum mainnet.

pub fn consensus_type(&self) -> ConsensusType

Retrieves the consensus type for the specified hardfork.

pub fn is_proof_of_stake(&self) -> bool

Checks if the hardfork uses Proof of Stake consensus.

pub fn is_proof_of_work(&self) -> bool

Checks if the hardfork uses Proof of Work consensus.

pub fn activation_timestamp(&self, chain: Chain) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the given chain.

pub fn mainnet_activation_timestamp(&self) -> Option<u64>

Retrieves the activation timestamp for the specified hardfork on the Ethereum mainnet.